@@ -8,7 +8,7 @@
     "docs": "jsdoc -c jsDocsConfig.json -r -t ./node_modules/better-docs",
     "lint": "eslint lib/",
     "test": "jest",
-    "postinstall": "node postinstall.js"
+    "postinstall": "node postinstall.mjs"
   },
   "repository": {
     "type": "git",
@@ -20,11 +20,11 @@
     "clans"
   ],
   "dependencies": {
-    "@microsoft/signalr": "^8.0.0",
-    "chalk": "^4.0.0",
+    "@microsoft/signalr": "^8.0.7",
+    "chalk": "^5.3.0",
     "cheerio": "^1.0.0-rc.10",
-    "entities": "^4.3.0",
-    "figlet": "^1.4.0",
+    "entities": "^5.0.0",
+    "figlet": "^1.7.0",
     "postman-request": "^2.88.1-postman.34",
     "signalr-client": "0.0.20"
   },
@@ -40,22 +40,23 @@
   ],
   "homepage": "https://github.com/noblox/noblox.js",
   "devDependencies": {
-    "@auto-it/conventional-commits": "^10.30.0",
-    "auto": "^10.30.0",
-    "better-docs": "2.7.2",
-    "dotenv": "16.0.3",
-    "eslint": "^7.28.0",
-    "eslint-config-standard": "^16.0.3",
-    "eslint-plugin-import": "^2.25.3",
-    "eslint-plugin-jest": "27.2.2",
-    "eslint-plugin-node": "^11.0.0",
-    "eslint-plugin-promise": "^5.1.0",
-    "jest": "^29.5.0",
-    "jest-extended": "^3.2.4",
-    "jsdoc": "3.6.10",
+    "@auto-it/conventional-commits": "^11.2.0",
     "JSONStream": "^1.3.1",
+    "auto": "^11.2.0",
+    "better-docs": "2.7.3",
+    "dotenv": "16.4.5",
+    "eslint": "^9.8.0",
+    "eslint-config-standard": "^17.1.0",
+    "eslint-plugin-import": "^2.29.1",
+    "eslint-plugin-jest": "28.6.0",
+    "eslint-plugin-n": "^17.10.1",
+    "eslint-plugin-node": "^11.0.0",
+    "eslint-plugin-promise": "^7.0.0",
+    "jest": "^29.7.0",
+    "jest-extended": "^4.0.2",
+    "jsdoc": "4.0.3",
     "progress": "^2.0.0",
-    "standard": "^17.0.0"
+    "standard": "^17.1.0"
   },
   "jest": {
     "setupFiles": [
@@ -83,6 +84,7 @@
     "onlyPublishWithReleaseLabel": true
   },
   "engines": {
-    "node": ">=16"
-  }
+    "node": ">=18.18"
+  },
+  "packageManager": "yarn@1.22.19+sha1.4ba7fc5c6e704fce2066ecbfb0b0d8976fe62447"
 }
