@@ -34,17 +34,17 @@ exports.func = function (args) {
   const onMessage = new events.EventEmitter()
   let waitingForRequest = false
   let latest
-  getMessages({ jar: jar, messageTab: 'Inbox', pageNumber: 0, pageSize: 1 })
+  getMessages({ jar, messageTab: 'Inbox', pageNumber: 0, pageSize: 1 })
     .then(function (initial) {
       latest = initial.collection[0] ? initial.collection[0].id : 0
-      const notifications = onNotification({ jar: jar })
+      const notifications = onNotification({ jar })
       notifications.on('data', function (name, message) {
         if (name === 'MessageNotification' && message.Type === 'Created') {
           if (waitingForRequest) {
             waitingForRequest = false
           } else {
             getMessages({
-              jar: jar,
+              jar,
               messageTab: 'Inbox',
               pageNumber: 0,
               pageSize: 1
